@@ -15,40 +15,42 @@
     "node": "4.2.x"
   },
   "dependencies": {
-    "autoprefixer": "^6.0.2",
-    "babel": "^5.0.8",
-    "babel-core": "^5.0.8",
-    "babel-eslint": "^4.1.2",
-    "babel-loader": "^5.0.0",
+    "autoprefixer": "^6.1.0",
+    "babel": "^5.8.29",
+    "babel-core": "^5.8.29",
+    "babel-eslint": "^4.1.5",
+    "babel-loader": "^5.3.3",
     "babel-plugin-react-transform": "^1.1.1",
     "bluebird": "^3.0.5",
     "body-parser": "^1.12.4",
-    "chai": "^3.0.0",
-    "chai-immutable": "^1.1.1",
+    "chai": "^3.4.1",
+    "chai-immutable": "^1.5.2",
     "classnames": "^2.1.1",
     "compression": "^1.4.0",
     "cors": "^2.6.0",
     "css-loader": "^0.22.0",
-    "eslint": "^1.4.1",
-    "eslint-plugin-react": "^3.3.2",
-    "express": "^4.11.2",
+    "del": "^2.1.0",
+    "eslint": "1.9.0",
+    "eslint-plugin-import": "^0.10.0",
+    "eslint-plugin-react": "^3.8.0",
+    "estraverse": "^4.1.1",
+    "express": "^4.13.3",
     "express-device": "^0.4.1",
     "extract-text-webpack-plugin": "^0.9.1",
     "file-loader": "^0.8.1",
     "glob": "^5.0.13",
     "gulp": "^3.9.0",
     "gulp-bg": "0.0.7",
-    "gulp-eslint": "^1.0.0",
+    "gulp-eslint": "^1.1.0",
     "gulp-mocha": "^2.1.3",
     "gulp-shell": "^0.5.0",
     "gulp-util": "^3.0.7",
-    "hash-file": "^2.1.0",
     "history": "^1.9.0",
     "immutable": "^3.6.4",
     "intl": "^1.0.0",
     "intl-messageformat": "^1.1.0",
     "intl-relativeformat": "^1.1.0",
-    "invariant": "^2.1.0",
+    "invariant": "^2.1.2",
     "ip": "^1.0.2",
     "isomorphic-fetch": "^2.1.1",
     "jsdom": "^7.0.2",
@@ -56,32 +58,34 @@
     "less-loader": "^2.0.0",
     "nconf": "^0.8.1",
     "node-sass": "^3.1.2",
-    "nodemon": "<1.8.0",
+    "nodemon": "^1.8.1",
     "normalize.css": "^3.0.2",
     "postcss-loader": "^0.7.0",
     "react": "^0.14.2",
     "react-addons-test-utils": "^0.14.1",
-    "react-document-title": "^2.0.0",
     "react-dom": "^0.14.0",
+    "react-helmet": "^2.1.1",
     "react-intl": "^2.0.0-pr-1",
     "react-native": "^0.14.0",
     "react-native-browser-polyfill": "^0.1.1",
     "react-native-side-menu": "^0.14.2",
     "react-pure-render": "^1.0.1",
-    "react-redux": "^3.0.1",
-    "react-router": "^1.0.0-rc1",
+    "react-redux": "^3.0.0",
+    "react-router": "^1.0.0",
     "react-transform-catch-errors": "^1.0.0",
     "react-transform-hmr": "^1.0.1",
     "redbox-react": "^1.0.5",
     "redux": "^3.0.4",
     "redux-logger": "^2.0.1",
-    "redux-promise-middleware": "^2.1.1",
+    "redux-promise-middleware": "2.2.3",
     "redux-storage": "^1.2.4",
+    "regenerator": "^0.8.42",
     "run-sequence": "^1.0.2",
     "sass-loader": "^3.0.0",
     "serialize-javascript": "^1.1.2",
     "serve-favicon": "^2.2.0",
     "sinon": "^1.17.2",
+    "sinon-as-promised": "^4.0.0",
     "sinon-chai": "^2.8.0",
     "style-loader": "^0.13.0",
     "stylus-loader": "^1.0.0",
