@@ -1,5 +1,4 @@
-const hashFile = require('hash-file');
-const nconf = require('nconf');
+import nconf from 'nconf';
 
 const isProduction = process.env.NODE_ENV === 'production';
 
@@ -7,22 +6,8 @@ const isProduction = process.env.NODE_ENV === 'production';
 // to production server.
 nconf.env('__');
 
-function getAssetHash(filePath) {
-  if (!isProduction) return '';
-  try {
-    return hashFile.sync(filePath);
-  }
-  catch (e) {
-    return '';
-  }
-}
-
 // Remember, never put production secrets in config. Use nconf.
 const config = {
-  assetsHashes: {
-    appCss: getAssetHash('build/app.css'),
-    appJs: getAssetHash('build/app.js')
-  },
   isProduction: isProduction,
   googleAnalyticsId: 'UA-XXXXXXX-X',
   port: process.env.PORT || 8001,
@@ -33,4 +18,4 @@ const config = {
 // like loading config from json and more. Check out nconf docs.
 nconf.defaults(config);
 
-module.exports = nconf.get();
+export default nconf.get();
