@@ -1,4 +1,5 @@
 import Component from 'react-pure-render/component';
+import Helmet from 'react-helmet';
 import React, {PropTypes} from 'react';
 import focusInvalidField from '../lib/focusInvalidField';
 
@@ -15,16 +16,14 @@ export default class Login extends Component {
     msg: PropTypes.object.isRequired
   }
 
-  onFormSubmit(e) {
+  async onFormSubmit(e) {
     e.preventDefault();
-    const {actions, auth: {form}} = this.props;
-    actions.login(form.fields)
-      .then(({error, payload: maybeValidationError}) => {
-        if (error)
-          focusInvalidField(this, maybeValidationError);
-        else
-          this.redirectAfterLogin();
-      });
+    const {actions: {login}, auth: {form}} = this.props;
+    const {error, payload: maybeValidationError} = await login(form.fields);
+    if (error)
+      focusInvalidField(this, maybeValidationError);
+    else
+      this.redirectAfterLogin();
   }
 
   // TODO: Use redux-react-router.
@@ -41,6 +40,7 @@ export default class Login extends Component {
 
     return (
       <div className="login">
+        <Helmet title="Login" />
         <form onSubmit={e => this.onFormSubmit(e)}>
           <fieldset disabled={form.disabled}>
             <legend>{msg.legend}</legend>
