@@ -1,6 +1,6 @@
 import AuthLogin from './Login.react';
 import Component from 'react-pure-render/component';
-import DocumentTitle from 'react-document-title';
+import Helmet from 'react-helmet';
 import React, {PropTypes} from 'react';
 
 export default class Page extends Component {
@@ -13,11 +13,10 @@ export default class Page extends Component {
     const {msg} = this.props;
 
     return (
-      <DocumentTitle title={msg.auth.login.title}>
-        <div className="login-page">
-          <AuthLogin {...this.props} />
-        </div>
-      </DocumentTitle>
+      <div className="login-page">
+        <Helmet title={msg.auth.login.title} />
+        <AuthLogin {...this.props} />
+      </div>
     );
   }
 
