@@ -1,5 +1,5 @@
 import Component from 'react-pure-render/component';
-import DocumentTitle from 'react-document-title';
+import Helmet from 'react-helmet';
 import React, {PropTypes} from 'react';
 import {FormattedHTMLMessage} from 'react-intl';
 import {Link} from 'react-router';
@@ -16,30 +16,30 @@ export default class Page extends Component {
     const {msg: {home: msg}} = this.props;
 
     return (
-      <DocumentTitle title={msg.title}>
-        <div className="home-page">
-          <p>
-            <FormattedHTMLMessage defaultMessage={msg.infoHtml} />
-          </p>
-          <div className="tocheck">
-            <h2>{msg.toCheck.h2}</h2>
-            <ul>
-              {msg.toCheck.list.map(({key, text}) =>
-                <li key={key}>
-                  <FormattedHTMLMessage defaultMessage={text} />
-                </li>
-              )}
-              <li>
-                {msg.toCheck.isomorphicPage}{' '}
-                <Link to="/this-is-not-the-web-page-you-are-looking-for">404</Link>
+      <div className="home-page">
+        <Helmet title={msg.title} />
+        <p>
+          <FormattedHTMLMessage defaultMessage={msg.infoHtml} />
+        </p>
+        <div className="tocheck">
+          <h2>{msg.toCheck.h2}</h2>
+          <img alt="50x50 placeholder" src="/assets/img/50x50.png"/>
+          <ul>
+            {msg.toCheck.list.map(({key, text}) =>
+              <li key={key}>
+                <FormattedHTMLMessage defaultMessage={text} />
               </li>
-              <li>
-                {msg.toCheck.andMuchMore}
-              </li>
-            </ul>
-          </div>
+            )}
+            <li>
+              {msg.toCheck.isomorphicPage}{' '}
+              <Link to="/this-is-not-the-web-page-you-are-looking-for">404</Link>
+            </li>
+            <li>
+              {msg.toCheck.andMuchMore}
+            </li>
+          </ul>
         </div>
-      </DocumentTitle>
+      </div>
     );
   }
 
