@@ -1,6 +1,6 @@
 import AuthLogout from '../auth/Logout.react';
 import Component from 'react-pure-render/component';
-import DocumentTitle from 'react-document-title';
+import Helmet from 'react-helmet';
 import React, {PropTypes} from 'react';
 import RouterHandler from '../../common/components/RouterHandler.react';
 import {FormattedMessage} from 'react-intl';
@@ -21,25 +21,24 @@ export default class Page extends Component {
     const {children, msg, users: {viewer: {email}}} = this.props;
 
     return (
-      <DocumentTitle title={msg.me.title}>
-        <div className="me-page">
-          <ul>
-            <li><Link activeClassName="active" to="/me/profile">{msg.profile.title}</Link></li>
-            <li><Link activeClassName="active" to="/me/settings">{msg.settings.title}</Link></li>
-          </ul>
-          {children
-            ? <RouterHandler {...this.props} />
-            : <p>
-                <FormattedMessage
-                  defaultMessage={msg.me.welcome}
-                  id={'msg.me.welcome'}
-                  values={{email}}
-                />
-              </p>
-          }
-          <AuthLogout msg={msg.auth.logout} />
-        </div>
-      </DocumentTitle>
+      <div className="me-page">
+        <Helmet title={msg.me.title} />
+        <ul>
+          <li><Link activeClassName="active" to="/me/profile">{msg.profile.title}</Link></li>
+          <li><Link activeClassName="active" to="/me/settings">{msg.settings.title}</Link></li>
+        </ul>
+        {children
+          ? <RouterHandler {...this.props} />
+          : <p>
+              <FormattedMessage
+                defaultMessage={msg.me.welcome}
+                id={'msg.me.welcome'}
+                values={{email}}
+              />
+            </p>
+        }
+        <AuthLogout msg={msg.auth.logout} />
+      </div>
     );
   }
 
