@@ -1,5 +1,5 @@
 import Component from 'react-pure-render/component';
-import DocumentTitle from 'react-document-title';
+import Helmet from 'react-helmet';
 import React, {PropTypes} from 'react';
 import {Link} from 'react-router';
 
@@ -13,13 +13,12 @@ export default class NotFound extends Component {
     const {msg: {notFound: msg}} = this.props;
 
     return (
-      <DocumentTitle title={msg.title}>
-        <div className="notfound-page">
-          <h1>{msg.header}</h1>
-          <p>{msg.message}</p>
-          <Link to="/">{msg.continueMessage}</Link>
-        </div>
-      </DocumentTitle>
+      <div className="notfound-page">
+        <Helmet title={msg.title} />
+        <h1>{msg.header}</h1>
+        <p>{msg.message}</p>
+        <Link to="/">{msg.continueMessage}</Link>
+      </div>
     );
   }
 
