@@ -1,6 +1,6 @@
 import Buttons from './Buttons.react';
 import Component from 'react-pure-render/component';
-import DocumentTitle from 'react-document-title';
+import Helmet from 'react-helmet';
 import List from './List.react';
 import NewTodo from './NewTodo.react';
 import React, {PropTypes} from 'react';
@@ -22,13 +22,12 @@ export default class Page extends Component {
     const {actions, msg: {todos: msg}, todos: {newTodo, list}} = this.props;
 
     return (
-      <DocumentTitle title={msg.title}>
-        <div className="todos-page">
-          <NewTodo {...{actions, msg, newTodo}} />
-          <List {...{actions, list, msg}} />
-          <Buttons clearAllEnabled={list.size > 0} {...{actions, msg}} />
-        </div>
-      </DocumentTitle>
+      <div className="todos-page">
+        <Helmet title={msg.title} />
+        <NewTodo {...{actions, msg, newTodo}} />
+        <List {...{actions, list, msg}} />
+        <Buttons clearAllEnabled={list.size > 0} {...{actions, msg}} />
+      </div>
     );
   }
 
