@@ -12,8 +12,10 @@ app.use(compression());
 
 // app.use(favicon('assets/img/favicon.ico'));
 
+
 // Serve the static assets. We can cache them as they include hashes.
-app.use('/assets/img', express.static('../assets/img', {maxAge: '200d'}));
+// express.static is relative to the directory where you launch your node process
+app.use('/assets/img', express.static('assets/img', {maxAge: '200d'}));
 app.use('/_assets', express.static('build', {maxAge: '200d'}));
 
 // Intl.
