@@ -1,9 +1,10 @@
-import DocumentTitle from 'react-document-title';
+import Helmet from 'react-helmet';
 import Html from './Html.react';
 import Promise from 'bluebird';
 import React from 'react';
 import ReactDOMServer from 'react-dom/server';
 import config from '../config';
+import getAppAssetFilenamesAsync from './assets';
 import configureStore from '../../common/configureStore';
 import createRoutes from '../../browser/createRoutes';
 import serialize from 'serialize-javascript';
@@ -28,7 +29,7 @@ export default function render(req, res, next) {
   const routes = createRoutes(() => store.getState());
   const location = createMemoryHistory().createLocation(req.url);
 
-  match({routes, location}, (error, redirectLocation, renderProps) => {
+  match({routes, location}, async (error, redirectLocation, renderProps) => {
 
     if (redirectLocation) {
       res.redirect(301, redirectLocation.pathname + redirectLocation.search);
@@ -46,14 +47,18 @@ export default function render(req, res, next) {
     //   return;
     // }
 
-    fetchComponentData(store.dispatch, req, renderProps)
-      .then(() => renderPage(store, renderProps, req))
-      .then(html => res.send(html))
-      .catch(next);
+    try {
+      await fetchComponentDataAsync(store.dispatch, renderProps);
+      const html = await renderPageAsync(store, renderProps, req);
+      res.send(html);
+    }
+    catch (e) {
+      next(e);
+    }
   });
 }
 
-function fetchComponentData(dispatch, req, {components, location, params}) {
+async function fetchComponentDataAsync(dispatch, {components, location, params}) { // eslint-disable-line space-before-function-paren
   const fetchActions = components.reduce((actions, component) => {
     return actions.concat(component.fetchActions || []);
   }, []);
@@ -64,27 +69,27 @@ function fetchComponentData(dispatch, req, {components, location, params}) {
   // Because redux-promise-middleware always returns fulfilled promise, we have
   // to detect errors manually.
   // https://github.com/pburtchaell/redux-promise-middleware#usage
-  return Promise.all(promises).then(results => {
-    results.forEach(result => {
-      if (result.error)
-        throw result.payload;
-    });
+  const results = await Promise.all(promises);
+  results.forEach(result => {
+    if (result.error)
+      throw result.payload;
   });
 }
 
-function renderPage(store, renderProps, req) {
+async function renderPageAsync(store, renderProps, req) { // eslint-disable-line space-before-function-paren
   const clientState = store.getState();
   const {headers, hostname} = req;
   const appHtml = getAppHtml(store, renderProps);
-  const scriptHtml = getScriptHtml(clientState, headers, hostname);
+  const {js: appJsFilename, css: appCssFilename} = await getAppAssetFilenamesCachedAsync();
+  const scriptHtml = getScriptHtml(clientState, headers, hostname, appJsFilename);
 
   return '<!DOCTYPE html>' + ReactDOMServer.renderToStaticMarkup(
     <Html
-      appCssHash={config.assetsHashes.appCss}
+      appCssFilename={appCssFilename}
       bodyHtml={`<div id="app">${appHtml}</div>${scriptHtml}`}
       googleAnalyticsId={config.googleAnalyticsId}
+      helmet={Helmet.rewind()}
       isProduction={config.isProduction}
-      title={DocumentTitle.rewind()}
     />
   );
 }
@@ -99,7 +104,17 @@ function getAppHtml(store, renderProps) {
   );
 }
 
-function getScriptHtml(clientState, headers, hostname) {
+let appAssetFilenameCache = null;
+
+async function getAppAssetFilenamesCachedAsync() { // eslint-disable-line space-before-function-paren
+  if (appAssetFilenameCache) return appAssetFilenameCache;
+
+  appAssetFilenameCache = await getAppAssetFilenamesAsync();
+
+  return appAssetFilenameCache;
+}
+
+function getScriptHtml(clientState, headers, hostname, appJsFilename) {
   let scriptHtml = '';
 
   const ua = useragent.is(headers['user-agent']);
@@ -112,7 +127,7 @@ function getScriptHtml(clientState, headers, hostname) {
   }
 
   const appScriptSrc = config.isProduction
-    ? '/_assets/app.js?' + config.assetsHashes.appJs
+    ? `/_assets/${appJsFilename}`
     : `//${hostname}:${HOT_RELOAD_PORT}/build/app.js`;
 
   // Note how clientState is serialized. JSON.stringify is anti-pattern.
