@@ -1,3 +1,5 @@
+require('babel/register')({optional: ['es7']});
+
 const serverConfig = require('./config');
 
 if (!process.env.NODE_ENV)
@@ -15,11 +17,9 @@ else {
   global.Intl = require('intl');
 }
 
-require('babel/register')({optional: ['es7']});
-
 // To ignore webpack custom loaders on server.
-serverConfig.webpackStylesExtensions.forEach(function(ext) {
-  require.extensions['.' + ext] = function() {};
+serverConfig.webpackStylesExtensions.forEach(ext => {
+  require.extensions['.' + ext] = () => {};
 });
 
 require('./main');
