@@ -3,19 +3,39 @@
 import gulp from 'gulp';
 import gutil from 'gulp-util';
 import mocha from 'gulp-mocha';
+import path from 'path';
 
 function reportError(errorReporter) {
-  return errorReporter === 'process' ? process.exit.bind(process, 1) : gutil.log
+  return errorReporter === 'process' ? process.exit.bind(process, 1) : gutil.log;
 }
 
 export default function mochaRunCreator(errorReporter = 'process') {
-  return () => {
-    // read: false, not to load file contents
-    gulp.src('src/**/__test__/**/*.js', {read: false})
+  return (file) => {
+    let source = 'src/**/__test__/**/*.js';
+
+    if (file) {
+      // Do not run tests when changed something not JS
+      if (!/\.(js|jsx)?$/.test(file.path)) {
+        console.log(`Change happend on '${file.path}' but it is not valid JS file`); // eslint-disable-line no-console
+        return null;
+      }
+
+      if (file.path.indexOf('__test__') !== -1)
+        source = file.path;
+      else {
+        const parts = file.path.split(path.sep);
+        const filename = parts.pop(1);
+        const dir = parts.join(path.sep);
+        source = `${dir}/__test__/${filename.split('.')[0]}*.js`;
+      }
+    }
+
+    console.log(`Running ${source}`); // eslint-disable-line no-console
+    gulp.src(source, {read: false})
       .pipe(mocha({
         require: ['./test/mochaSetup.js'],
         reporter: 'spec'
       }))
       .on('error', reportError(errorReporter));
-  }
+  };
 }
