@@ -1,6 +1,7 @@
 import path from 'path';
 
-const ABSOLUTE_BASE = path.normalize(path.join(__dirname, '..'));
+export const ABSOLUTE_BASE = path.normalize(path.join(__dirname, '..'));
+
 
 const constants = Object.freeze({
   ABSOLUTE_BASE: ABSOLUTE_BASE,
@@ -9,7 +10,18 @@ const constants = Object.freeze({
   DIST_DIR: path.join(ABSOLUTE_BASE, 'dist'),
   SRC_DIR: path.join(ABSOLUTE_BASE, 'src'),
   ASSETS_DIR: path.join(ABSOLUTE_BASE, 'assets'),
+<<<<<<< HEAD
   HOT_RELOAD_PORT: 8081
+=======
+  HOT_RELOAD_PORT: process.env.HOT_RELOAD_PORT || 8080,
+>>>>>>> upstream/master
 });
+
+export const NODE_MODULES_DIR = constants.NODE_MODULES_DIR;
+export const BUILD_DIR = constants.BUILD_DIR;
+export const DIST_DIR = constants.DIST_DIR;
+export const SRC_DIR = constants.SRC_DIR;
+export const ASSETS_DIR = constants.ASSETS_DIR;
+export const HOT_RELOAD_PORT = constants.HOT_RELOAD_PORT;
 
 export default constants;
