@@ -10,6 +10,7 @@ const webpackConfig = makeWebpackConfig(true);
 const compiler = webpack(webpackConfig);
 
 app.use(webpackDev(compiler, {
+  headers: {'Access-Control-Allow-Origin': '*'},
   noInfo: true,
   publicPath: webpackConfig.output.publicPath
 }));
@@ -17,5 +18,5 @@ app.use(webpackDev(compiler, {
 app.use(webpackHot(compiler));
 
 app.listen(webpackConfig.hotPort, () => {
-  console.log('Hot server started at port %s', webpackConfig.hotPort); // eslint-disable-line no-console
+  console.log('Hot server started at port %d', webpackConfig.hotPort); // eslint-disable-line no-console
 });
